@@ -31,79 +31,82 @@
     "stylelint": "stylelint \"**/*.css\"",
     "stylelint:fix": "stylelint \"**/*.css\" --fix",
     "test": "echo 'Run `npm run test-all` to run `npm test` for every package'",
-    "test-all": "lerna run test --stream",
+    "test-all": "lerna run test --concurrency=2 --stream",
     "test:backstop": "gulp backstop",
     "test:backstop:approve": "gulp backstop:approve",
     "test:backstop:reference": "gulp backstop:reference",
     "updated": "lerna updated",
     "watch": "gulp watch",
-    "install-husky": "node node_modules/husky/bin/install",
-    "commitmsg": "node scripts/validateCommitMessage.js",
     "release-notes": "node scripts/printReleaseNotesTemplate.js",
     "versions": "node scripts/normalizeDependencyVersions.js",
     "normalize-pkgfields": "node scripts/normalizePackageFields.js",
     "prettify": "lerna run clean && prettier --write \"**/*.js\""
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.0",
-    "@babel/core": "^7.1.0",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/plugin-transform-flow-comments": "^7.0.0",
-    "@babel/plugin-transform-flow-strip-types": "^7.0.0",
-    "@babel/polyfill": "^7.0.0",
-    "@babel/preset-env": "^7.1.0",
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-transform-flow-comments": "^7.4.4",
+    "@babel/plugin-transform-flow-strip-types": "^7.4.4",
+    "@babel/preset-env": "^7.4.5",
     "@babel/preset-flow": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.0.0",
-    "babel-eslint": "^9.0.0",
-    "babel-jest": "^23.6.0",
-    "babel-loader": "^8.0.0",
-    "babel-plugin-css-modules-transform": "^1.6.1",
-    "babel-plugin-istanbul": "^5.0.1",
+    "@babel/register": "^7.4.4",
+    "babel-eslint": "^10.0.2",
+    "babel-jest": "^24.8.0",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-css-modules-transform": "^1.6.2",
+    "babel-plugin-istanbul": "^5.1.4",
     "babel-plugin-lodash": "^3.3.4",
-    "backstopjs": "^3.8.5",
-    "boxen": "^1.3.0",
+    "backstopjs": "^4.1.9",
+    "boxen": "^4.1.0",
     "chalk": "^2.3.0",
-    "deepmerge": "^2.0.1",
-    "eslint": "^4.19.0",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-import-resolver-webpack": "^0.8.4",
+    "deepmerge": "^3.3.0",
+    "eslint": "^6.0.1",
+    "eslint-config-prettier": "^5.1.0",
+    "eslint-import-resolver-webpack": "^0.11.1",
     "eslint-plugin-es5": "^1.2.0",
-    "eslint-plugin-flowtype": "^2.46.1",
-    "eslint-plugin-flowtype-errors": "^3.5.1",
-    "eslint-plugin-import": "^2.9.0",
-    "eslint-plugin-prettier": "^2.6.0",
-    "eslint-plugin-react": "^7.7.0",
-    "execa": "^0.10.0",
-    "flow-bin": "^0.68.0",
+    "eslint-plugin-flowtype": "^3.11.1",
+    "eslint-plugin-flowtype-errors": "^4.1.0",
+    "eslint-plugin-import": "^2.18.0",
+    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-react": "^7.14.2",
+    "execa": "^1.0.0",
+    "flow-bin": "^0.101.1",
     "glob": "^7.1.2",
-    "globby": "^8.0.1",
-    "gulp": "^3.9.1",
+    "globby": "^9.2.0",
+    "gulp": "^4.0.2",
     "gulp-babel": "^8.0.0",
-    "gulp-filter": "^5.1.0",
+    "gulp-changed": "^4.0.0",
+    "gulp-filter": "^6.0.0",
     "gulp-newer": "^1.4.0",
     "gulp-plumber": "^1.2.0",
     "gulp-sourcemaps": "^2.6.4",
     "gulp-typescript": "^5.0.0",
     "gulp-util": "^3.0.8",
     "gulp-watch": "^5.0.1",
-    "hard-rejection": "^1.0.0",
-    "husky": "^0.14.3",
+    "hard-rejection": "^2.1.0",
+    "husky": "^2.5.0",
     "kill-port": "^1.3.2",
     "lerna": "^3.13.1",
-    "merge-stream": "^1.0.1",
+    "merge-stream": "^2.0.0",
     "minimist": "^1.2.0",
     "path-exists-cli": "^1.0.0",
     "pre-commit": "^1.2.2",
     "prettier": "^1.13.0",
     "rimraf": "^2.6.2",
-    "semver": "^5.5.0",
-    "stylelint": "^8.4.0",
+    "semver": "^6.1.2",
+    "stylelint": "^10.1.0",
     "stylelint-config-css-modules": "^1.1.0",
     "stylelint-config-standard": "^18.0.0",
-    "through2": "^2.0.3",
+    "through2": "^3.0.1",
     "typescript": "^3.2.2",
     "wait-port": "^0.2.2",
     "yarn": "^1.3.2"
+  },
+  "husky": {
+    "hooks": {
+      "commitmsg": "node scripts/validateCommitMessage.js"
+    }
   }
 }
