@@ -1,6 +1,6 @@
 {
   "name": "eslint-config-sanity",
-  "version": "0.140.0",
+  "version": "0.141.5",
   "description": "Shared eslint configs for sanity-io",
   "main": "index.js",
   "scripts": {
@@ -28,12 +28,12 @@
   },
   "homepage": "https://www.sanity.io/",
   "peerDependencies": {
-    "eslint": "^4.6"
+    "eslint": "^6.0.0"
   },
   "devDependencies": {
-    "eslint": "^4.19.0",
-    "eslint-plugin-import": "^2.9.0",
-    "eslint-plugin-react": "^7.7.0",
+    "eslint": "^6.0.1",
+    "eslint-plugin-import": "^2.18.0",
+    "eslint-plugin-react": "^7.14.2",
     "lodash.difference": "^4.5.0"
   }
 }
