@@ -2,13 +2,14 @@
 // Note: Node 8 compat, please!
 require('hard-rejection/register')
 
-const path = require('path')
+const fs = require('fs')
 const childProcess = require('child_process')
-const gulp = require('gulp')
+const path = require('path')
+const {series, parallel, src, dest, watch} = require('gulp')
 const newer = require('gulp-newer')
 const babel = require('gulp-babel')
-const watch = require('gulp-watch')
 const gutil = require('gulp-util')
+const changed = require('gulp-changed')
 const filter = require('gulp-filter')
 const plumber = require('gulp-plumber')
 const ts = require('gulp-typescript')
@@ -20,12 +21,21 @@ const mergeStream = require('merge-stream')
 const backstop = require('backstopjs')
 const waitPort = require('wait-port')
 const kill = require('kill-port')
+const {promisify} = require('util')
 
+const stat = promisify(fs.stat)
 const isWindows = /^win/.test(process.platform)
+
+const compareModified = async (stream, sourceFile, targetPath) => {
+  const targetStat = await stat(targetPath)
+
+  if (sourceFile.stat && Math.floor(sourceFile.stat.mtimeMs) > Math.floor(targetStat.mtimeMs)) {
+    stream.push(sourceFile)
+  }
+}
 
 const tsPaths = globby.sync(['./packages/@sanity/*/tsconfig.json'])
 const tsProjects = tsPaths.map(conf => ts.createProject(conf))
-const tsScripts = tsPaths.map(proj => `${path.dirname(proj)}/src/**/*.ts`)
 const scripts = [
   './packages/@sanity/*/src/**/*.js',
   './packages/sanity-plugin-*/src/**/*.js',
@@ -35,7 +45,7 @@ const scripts = [
 const assets = [
   './packages/@sanity/*/src/**/*',
   './packages/sanity-plugin-*/src/**/*',
-  './packages/groq/src/**/*.js'
+  './packages/groq/src/**/*'
 ]
 
 const srcOpts = {base: 'packages'}
@@ -64,8 +74,8 @@ if (path.win32 === path) {
   libFragment = '$1/lib/'
 }
 
-const logCompile = (file, enc, cb) => {
-  gutil.log('Compiling', `'${chalk.cyan(file._path || file.path)}'...`)
+const logCompile = from => (file, enc, cb) => {
+  gutil.log(`[${from}] Compiling`, `'${chalk.cyan(file._path || file.path)}'...`)
   cb(null, file)
 }
 
@@ -74,21 +84,48 @@ const mapToDest = orgPath => {
   return outPath
 }
 
-const dest = 'packages'
-
-gulp.task('default', ['build', 'build-ts'])
-
+const packagesPath = 'packages'
 const pkgPath = (cwd, sourcePath) => path.relative(path.join(cwd, 'packages'), sourcePath)
+const getLogErrorHandler = () => plumber({errorHandler: err => gutil.log(err.stack)})
+const watchJavaScript = series(buildJavaScript, function watchJS() {
+  watch(scripts, watchJavaScriptRebuild)
+})
 
-gulp.task('build', () => {
+const watchAssets = series(buildAssets, function watchAsset() {
+  watch(assets, buildAssets)
+})
+
+const watchTypeScript = series(buildTypeScript, function watchTS() {
+  tsProjects.forEach(project => {
+    const builder = function() {
+      return rebuildTypeScriptProject(project)
+    }
+
+    Object.defineProperty(builder, 'name', {
+      value: `buildTypeScript[${path.basename(project.projectDirectory)}]`
+    })
+
+    watch(`${project.projectDirectory}/src/**/*.ts`, builder)
+  })
+})
+
+exports.default = parallel(buildJavaScript, buildTypeScript)
+exports.build = parallel(buildJavaScript, buildTypeScript)
+exports.watchTypeScript = watchTypeScript
+exports.watch = parallel(watchJavaScript, watchTypeScript, watchAssets)
+
+function buildJavaScript() {
   const assetFilter = filter(['**/*.js'], {restore: true})
-
-  return gulp
-    .src(assets, srcOpts)
+  return src(assets, srcOpts)
     .pipe(plumber({errorHandler: err => gutil.log(err.stack)}))
-    .pipe(newer({map: mapToDest}))
     .pipe(assetFilter)
-    .pipe(through.obj(logCompile))
+    .pipe(
+      changed(packagesPath, {
+        transformPath: inp => mapToDest(inp),
+        hasChanged: compareModified
+      })
+    )
+    .pipe(through.obj(logCompile('JS')))
     .pipe(babel())
     .pipe(assetFilter.restore)
     .pipe(
@@ -98,12 +135,11 @@ gulp.task('build', () => {
         callback(null, file)
       })
     )
-    .pipe(gulp.dest(dest))
-})
+    .pipe(dest(packagesPath))
+}
 
-gulp.task('watch-js', () => {
-  return gulp
-    .src(scripts, srcOpts)
+function watchJavaScriptRebuild() {
+  return src(scripts, srcOpts)
     .pipe(plumber({errorHandler: err => gutil.log(err.stack)}))
     .pipe(
       through.obj((file, enc, callback) => {
@@ -112,54 +148,46 @@ gulp.task('watch-js', () => {
         callback(null, file)
       })
     )
-    .pipe(newer(dest))
-    .pipe(through.obj(logCompile))
+    .pipe(newer(packagesPath))
+    .pipe(through.obj(logCompile('JS')))
     .pipe(babel())
-    .pipe(gulp.dest(dest))
-})
+    .pipe(dest(packagesPath))
+}
 
-// build-ts / watch-ts-build
-;[
-  {name: 'build-ts', params: {throwOnError: true}},
-  {name: 'watch-ts-build', params: {throwOnError: false}}
-].forEach(task => {
-  gulp.task(task.name, () => {
-    const getErrorHandler = () =>
-      task.params.throwOnError
-        ? through.obj((chunk, enc, cb) => cb(null, chunk))
-        : plumber({errorHandler: err => gutil.log(err.stack)})
+function rebuildTypeScriptProject(project) {
+  return project
+    .src()
+    .pipe(through.obj(logCompile('TS')))
+    .pipe(getLogErrorHandler())
+    .pipe(project())
+    .js.pipe(getLogErrorHandler())
+    .pipe(dest(project.options.outDir))
+}
 
-    const streams = tsProjects.map(project => {
-      const src = project
+function buildTypeScript() {
+  return mergeStream(
+    ...tsProjects.map(project => {
+      const compilation = project
         .src()
-        .pipe(getErrorHandler())
-        .pipe(sourcemaps.init())
+        .pipe(through.obj(logCompile('TS')))
         .pipe(project())
 
       return mergeStream(
-        src.dts
-          .pipe(getErrorHandler())
+        compilation.dts
           .pipe(sourcemaps.write('.', {includeContent: false, sourceRoot: './'}))
-          .pipe(gulp.dest(project.options.outDir)),
+          .pipe(dest(project.options.outDir)),
 
-        src.js
-          .pipe(getErrorHandler())
-          .pipe(through.obj(logCompile))
+        compilation.js
           .pipe(sourcemaps.write('.', {includeContent: false, sourceRoot: './'}))
-          .pipe(gulp.dest(project.options.outDir))
+          .pipe(dest(project.options.outDir))
       )
     })
+  )
+}
 
-    return mergeStream(...streams)
-  })
-})
-
-gulp.task('watch-ts', ['watch-ts-build'], () => gulp.watch(tsScripts, ['watch-ts-build']))
-
-gulp.task('watch-assets', () => {
-  return gulp
-    .src(assets, srcOpts)
-    .pipe(filter(['**/*.*', '!**/*.js']))
+function buildAssets() {
+  return src(assets, srcOpts)
+    .pipe(filter(['**/*.*', '!**/*.js', '!**/*.ts']))
     .pipe(plumber({errorHandler: err => gutil.log(err.stack)}))
     .pipe(
       through.obj((file, enc, callback) => {
@@ -168,25 +196,15 @@ gulp.task('watch-assets', () => {
         callback(null, file)
       })
     )
-    .pipe(newer(dest))
+    .pipe(newer(packagesPath))
     .pipe(
       through.obj((file, enc, callback) => {
         gutil.log('Copying  ', `'${chalk.green(pkgPath(file.cwd, file._path))}'...`)
         callback(null, file)
       })
     )
-    .pipe(gulp.dest(dest))
-})
-
-gulp.task('watch', ['watch-js', 'watch-ts', 'watch-assets'], callback => {
-  watch(scripts, {debounceDelay: 200}, () => {
-    gulp.start('watch-js')
-  })
-
-  watch(assets, {debounceDelay: 200}, () => {
-    gulp.start('watch-assets')
-  })
-})
+    .pipe(dest(packagesPath))
+}
 
 const STUDIOS = [
   {name: 'test-studio', port: '3333'},
@@ -199,55 +217,43 @@ const STUDIOS = [
 ]
 
 STUDIOS.forEach(studio => {
-  gulp.task(studio.name, ['watch-js', 'watch-ts', 'watch-assets'], cb => {
-    watch(scripts, {debounceDelay: 200}, () => {
-      gulp.start('watch-js')
-    })
+  exports[studio.name] = series(
+    parallel(buildJavaScript, buildTypeScript, buildAssets),
+    parallel(exports.watch, function runStudio(cb) {
+      const projectPath = path.join(__dirname, 'packages', studio.name)
+      const proc = childProcess.spawn(
+        'sanity',
+        ['start', '--host', '0.0.0.0', '--port', studio.port],
+        {
+          shell: isWindows,
+          cwd: projectPath,
+          env: getProjectEnv(projectPath)
+        }
+      )
 
-    watch(assets, {debounceDelay: 200}, () => {
-      gulp.start('watch-assets')
+      proc.stdout.pipe(process.stdout)
+      proc.stderr.pipe(process.stderr)
     })
+  )
+})
 
-    const projectPath = path.join(__dirname, 'packages', studio.name)
-    const proc = childProcess.spawn(
-      'sanity',
-      ['start', '--host', '0.0.0.0', '--port', studio.port],
-      {
-        shell: isWindows,
-        cwd: projectPath,
-        env: getProjectEnv(projectPath)
-      }
-    )
+exports.storybook = series(
+  parallel(buildJavaScript, buildTypeScript, buildAssets),
+  parallel(exports.watch, function runStorybook(cb) {
+    const projectPath = path.join(__dirname, 'packages', 'storybook')
+    const proc = childProcess.spawn('npm', ['start'], {
+      shell: isWindows,
+      cwd: projectPath,
+      env: getProjectEnv(projectPath)
+    })
 
     proc.stdout.pipe(process.stdout)
     proc.stderr.pipe(process.stderr)
   })
-})
+)
 
-gulp.task('storybook', ['watch-js', 'watch-ts', 'watch-assets'], () => {
-  watch(scripts, {debounceDelay: 200}, () => {
-    gulp.start('watch-js')
-  })
-
-  watch(assets, {debounceDelay: 200}, () => {
-    gulp.start('watch-assets')
-  })
-
-  const projectPath = path.join(__dirname, 'packages', 'storybook')
-  const proc = childProcess.spawn('npm', ['start'], {
-    shell: isWindows,
-    cwd: projectPath,
-    env: getProjectEnv(projectPath)
-  })
-
-  proc.stdout.pipe(process.stdout)
-  proc.stderr.pipe(process.stderr)
-})
-
-gulp.task('backstop', cb => {
-  const {exec} = require('child_process')
-
-  exec('docker -v', (err, stdout, stderr) => {
+exports.backstop = function(cb) {
+  childProcess.exec('docker -v', err => {
     if (err) {
       throw new gutil.PluginError({
         plugin: 'backstop',
@@ -256,71 +262,73 @@ gulp.task('backstop', cb => {
     }
   })
 
-  gulp.start('backstop-test-studio')
-
   const params = {
     host: 'localhost',
     port: 5000,
     timeout: 100 * 60 * 10
   }
 
-  waitPort(params)
-    .then(open => {
-      if (open) {
-        backstop('test', {
-          docker: true,
-          config: './test/backstop/backstop.js'
-        })
-          .then(() => {
-            kill(params.port).then(() => {
-              gutil.log(gutil.colors.green('Backstop test success'))
-              // eslint-disable-next-line
-              process.exit(0)
-            })
-          })
-          .catch(() => {
-            kill(params.port)
-            throw new gutil.PluginError({
-              plugin: 'backstop',
-              message: 'Tests failed'
-            })
-          })
-      } else {
-        kill(params.port)
-        throw new gutil.PluginError({
-          plugin: 'backstop',
-          message: 'The backstop-studio did not start'
-        })
-      }
-    })
-    .catch(err => {
+  parallel(exports['backstop-test-studio'], async () => {
+    let open
+    try {
+      open = await waitPort(params)
+    } catch (err) {
       kill(params.port)
       throw new gutil.PluginError({
         plugin: 'backstop',
         message: `An unknown error occured while waiting for the port: ${err}`
       })
-    })
-})
+    }
 
-gulp.task('backstop:approve', cb => {
+    if (!open) {
+      kill(params.port)
+      throw new gutil.PluginError({
+        plugin: 'backstop',
+        message: 'The backstop-studio did not start'
+      })
+    }
+
+    try {
+      await backstop('test', {
+        docker: true,
+        config: './test/backstop/backstop.js'
+      })
+
+      await kill(params.port)
+      gutil.log(gutil.colors.green('Backstop test success'))
+      // eslint-disable-next-line no-process-exit
+      process.exit(0)
+    } catch (err) {
+      kill(params.port)
+      throw new gutil.PluginError({
+        plugin: 'backstop',
+        message: 'Tests failed'
+      })
+    }
+
+    cb()
+  })()
+}
+
+exports['backstop:approve'] = cb => {
   backstop('approve', {
     docker: true,
     config: './test/backstop/backstop.js'
   })
-})
+}
 
-gulp.task('backstop:reference', cb => {
-  backstop('reference', {
-    docker: true,
-    config: './test/backstop/backstop.js'
-  })
-    .then(() => {
-      gutil.log(gutil.colors.green('References created'))
+exports['backstop:reference'] = async cb => {
+  try {
+    await backstop('reference', {
+      docker: true,
+      config: './test/backstop/backstop.js'
     })
-    .catch(() => {
-      throw new gutil.PluginError({
-        plugin: 'backstop',
-        message: 'Making references failed'
-      })
+
+    gutil.log(gutil.colors.green('References created'))
+  } catch (err) {
+    throw new gutil.PluginError({
+      plugin: 'backstop',
+      message: 'Making references failed'
     })
-})
+  }
+}
