@@ -1,7 +1,6 @@
 module.exports = {
-  transform: {'^.+\\.jsx?$': 'babel-jest'},
-  testRegex: 'test\\/.*\\.test\\.js',
+  testRegex: 'test\\/.*\\.test\\.js$',
   testEnvironment: 'node',
-  setupTestFrameworkScriptFile: './test/setup.js',
+  setupFilesAfterEnv: ['./test/setup.js'],
   collectCoverageFrom: ['src/**.js']
 }
