@@ -10,15 +10,21 @@ module.exports = (config, overrides = {}) => {
     headers.Authorization = `Bearer ${token}`
   }
 
-  if (!config.useProjectHostname && config.projectId) {
+  if (!overrides.useGlobalApi && (!config.useProjectHostname && config.projectId)) {
     headers[projectHeader] = config.projectId
   }
+
+  const withCredentials = Boolean(
+    typeof overrides.withCredentials === 'undefined'
+      ? config.token || config.withCredentials
+      : overrides.withCredentials
+  )
 
   const timeout = typeof overrides.timeout === 'undefined' ? config.timeout : overrides.timeout
   return assign({}, overrides, {
     headers: assign({}, headers, overrides.headers || {}),
     timeout: typeof timeout === 'undefined' ? 5 * 60 * 1000 : timeout,
     json: true,
-    withCredentials: Boolean(config.token || config.withCredentials)
+    withCredentials
   })
 }
