@@ -1,4 +1,4 @@
 module.exports = {
   presets: ['@babel/flow'],
-  extends: '../../../.babelrc.js'
+  extends: '../../../.babelrc'
 }
