@@ -19,7 +19,8 @@ const INHERITED_FIELDS = [
   'jsonType',
   'description',
   'options',
-  'fieldsets'
+  'fieldsets',
+  'icon'
 ]
 
 const BLOCK_CORE = {
