@@ -2,7 +2,7 @@ import getConfig from './getConfig'
 import reduceConfig from './reduceConfig'
 import lazyRequire from './lazyRequire'
 import dynamicRequire from './dynamicRequire'
-import safeJson from './safeJson'
+import * as safeJson from './safeJson'
 import getSanityVersions from './getSanityVersions'
 import * as pathTools from './pathTools'
 
