@@ -1,4 +1,5 @@
 /* eslint-disable max-nested-callbacks */
+import {test} from 'tap'
 import {fromString, toString, get} from '../src/pathUtils'
 
 const srcObject = {
@@ -12,163 +13,180 @@ const srcObject = {
   ]
 }
 
-describe('path utilities', () => {
-  describe('fromString', () => {
-    test('throws if not a string', () => {
-      expect(() => fromString()).toThrowError('Path is not a string')
-      expect(() => fromString(13)).toThrowError('Path is not a string')
-      expect(() => fromString(null)).toThrowError('Path is not a string')
-      expect(() => fromString(false)).toThrowError('Path is not a string')
-    })
+test('fromString: throws if not a string', t => {
+  t.throws(() => fromString(), 'Path is not a string')
+  t.throws(() => fromString(13), 'Path is not a string')
+  t.throws(() => fromString(null), 'Path is not a string')
+  t.throws(() => fromString(false), 'Path is not a string')
+  t.end()
+})
 
-    test('handles plain property segments', () => {
-      expect(fromString('foo')).toEqual(['foo'])
-    })
+test('fromString: handles plain property segments', t => {
+  t.strictDeepEqual(fromString('foo'), ['foo'])
+  t.end()
+})
 
-    test('handles index segments', () => {
-      expect(fromString('[0]')).toEqual([0])
-      expect(fromString('[1337]')).toEqual([1337])
-    })
+test('fromString: handles index segments', t => {
+  t.strictDeepEqual(fromString('[0]'), [0])
+  t.strictDeepEqual(fromString('[1337]'), [1337])
+  t.end()
+})
 
-    test('handles key segments', () => {
-      expect(fromString('[_key == "foo"]')).toEqual([{_key: 'foo'}])
-      expect(fromString("[ _key== 'B4z']")).toEqual([{_key: 'B4z'}])
-      expect(fromString('[_key=="bar"]')).toEqual([{_key: 'bar'}])
-    })
+test('fromString: handles key segments', t => {
+  t.strictDeepEqual(fromString('[_key == "foo"]'), [{_key: 'foo'}])
+  t.strictDeepEqual(fromString("[ _key== 'B4z']"), [{_key: 'B4z'}])
+  t.strictDeepEqual(fromString('[_key=="bar"]'), [{_key: 'bar'}])
+  t.end()
+})
 
-    test('handles deep prop segments', () => {
-      expect(fromString('foo.bar')).toEqual(['foo', 'bar'])
-      expect(fromString('bar.foo')).toEqual(['bar', 'foo'])
-      expect(fromString('bar.foo.baz')).toEqual(['bar', 'foo', 'baz'])
-    })
+test('fromString: handles deep prop segments', t => {
+  t.strictDeepEqual(fromString('foo.bar'), ['foo', 'bar'])
+  t.strictDeepEqual(fromString('bar.foo'), ['bar', 'foo'])
+  t.strictDeepEqual(fromString('bar.foo.baz'), ['bar', 'foo', 'baz'])
+  t.end()
+})
 
-    test('handles deep array index segments', () => {
-      expect(fromString('foo[13]')).toEqual(['foo', 13])
-      expect(fromString('bar.foo[3]')).toEqual(['bar', 'foo', 3])
-      expect(fromString('[3][18]')).toEqual([3, 18])
-    })
+test('fromString: handles deep array index segments', t => {
+  t.strictDeepEqual(fromString('foo[13]'), ['foo', 13])
+  t.strictDeepEqual(fromString('bar.foo[3]'), ['bar', 'foo', 3])
+  t.strictDeepEqual(fromString('[3][18]'), [3, 18])
+  t.end()
+})
 
-    test('handles deep key segments', () => {
-      expect(fromString('foo[_key=="bar"].body[_key=="13ch"')).toEqual([
-        'foo',
-        {_key: 'bar'},
-        'body',
-        {_key: '13ch'}
-      ])
-      expect(fromString('bar.foo[3][_key == "seg"]')).toEqual(['bar', 'foo', 3, {_key: 'seg'}])
-      expect(fromString('[_key=="foo"][_key== "bar"][_key =="baz"][  _key   == "seg"  ')).toEqual([
-        {_key: 'foo'},
-        {_key: 'bar'},
-        {_key: 'baz'},
-        {_key: 'seg'}
-      ])
-    })
-  })
+test('fromString: handles deep key segments', t => {
+  t.strictDeepEqual(fromString('foo[_key=="bar"].body[_key=="13ch"'), [
+    'foo',
+    {_key: 'bar'},
+    'body',
+    {_key: '13ch'}
+  ])
+  t.strictDeepEqual(fromString('bar.foo[3][_key == "seg"]'), ['bar', 'foo', 3, {_key: 'seg'}])
+  t.strictDeepEqual(fromString('[_key=="foo"][_key== "bar"][_key =="baz"][  _key   == "seg"  '), [
+    {_key: 'foo'},
+    {_key: 'bar'},
+    {_key: 'baz'},
+    {_key: 'seg'}
+  ])
+  t.end()
+})
 
-  describe('toString', () => {
-    test('throws if not an array', () => {
-      expect(() => toString()).toThrowError('Path is not an array')
-      expect(() => toString(13)).toThrowError('Path is not an array')
-      expect(() => toString(null)).toThrowError('Path is not an array')
-      expect(() => toString(false)).toThrowError('Path is not an array')
-    })
+test('toString: throws if not an array', t => {
+  t.throws(() => toString(), 'Path is not an array')
+  t.throws(() => toString(13), 'Path is not an array')
+  t.throws(() => toString(null), 'Path is not an array')
+  t.throws(() => toString(false), 'Path is not an array')
+  t.end()
+})
 
-    test('handles plain property segments', () => {
-      expect(toString(['foo'])).toEqual('foo')
-    })
+test('toString: handles plain property segments', t => {
+  t.strictDeepEqual(toString(['foo']), 'foo')
+  t.end()
+})
 
-    test('handles index segments', () => {
-      expect(toString([0])).toEqual('[0]')
-      expect(toString([1337])).toEqual('[1337]')
-    })
+test('toString: handles index segments', t => {
+  t.strictDeepEqual(toString([0]), '[0]')
+  t.strictDeepEqual(toString([1337]), '[1337]')
+  t.end()
+})
 
-    test('handles key segments', () => {
-      expect(toString([{_key: 'foo'}])).toEqual('[_key=="foo"]')
-      expect(toString([{_key: 'B4z'}])).toEqual('[_key=="B4z"]')
-      expect(toString([{_key: 'bar'}])).toEqual('[_key=="bar"]')
-    })
+test('toString: handles key segments', t => {
+  t.strictDeepEqual(toString([{_key: 'foo'}]), '[_key=="foo"]')
+  t.strictDeepEqual(toString([{_key: 'B4z'}]), '[_key=="B4z"]')
+  t.strictDeepEqual(toString([{_key: 'bar'}]), '[_key=="bar"]')
+  t.end()
+})
 
-    test('handles deep prop segments', () => {
-      expect(toString(['foo', 'bar'])).toEqual('foo.bar')
-      expect(toString(['bar', 'foo'])).toEqual('bar.foo')
-      expect(toString(['bar', 'foo', 'baz'])).toEqual('bar.foo.baz')
-    })
+test('toString: handles deep prop segments', t => {
+  t.strictDeepEqual(toString(['foo', 'bar']), 'foo.bar')
+  t.strictDeepEqual(toString(['bar', 'foo']), 'bar.foo')
+  t.strictDeepEqual(toString(['bar', 'foo', 'baz']), 'bar.foo.baz')
+  t.end()
+})
 
-    test('handles deep array index segments', () => {
-      expect(toString(['foo', 13])).toEqual('foo[13]')
-      expect(toString(['bar', 'foo', 3])).toEqual('bar.foo[3]')
-      expect(toString([3, 18])).toEqual('[3][18]')
-    })
+test('toString: handles deep array index segments', t => {
+  t.strictDeepEqual(toString(['foo', 13]), 'foo[13]')
+  t.strictDeepEqual(toString(['bar', 'foo', 3]), 'bar.foo[3]')
+  t.strictDeepEqual(toString([3, 18]), '[3][18]')
+  t.end()
+})
 
-    test('handles deep key segments', () => {
-      expect(toString(['foo', {_key: 'bar'}, 'body', {_key: '13ch'}])).toEqual(
-        'foo[_key=="bar"].body[_key=="13ch"]'
-      )
-      expect(toString(['bar', 'foo', 3, {_key: 'seg'}])).toEqual('bar.foo[3][_key=="seg"]')
-      expect(toString([{_key: 'foo'}, {_key: 'bar'}, {_key: 'baz'}, {_key: 'seg'}])).toEqual(
-        '[_key=="foo"][_key=="bar"][_key=="baz"][_key=="seg"]'
-      )
-    })
+test('toString: handles deep key segments', t => {
+  t.strictDeepEqual(
+    toString(['foo', {_key: 'bar'}, 'body', {_key: '13ch'}]),
+    'foo[_key=="bar"].body[_key=="13ch"]'
+  )
+  t.strictDeepEqual(toString(['bar', 'foo', 3, {_key: 'seg'}]), 'bar.foo[3][_key=="seg"]')
+  t.strictDeepEqual(
+    toString([{_key: 'foo'}, {_key: 'bar'}, {_key: 'baz'}, {_key: 'seg'}]),
+    '[_key=="foo"][_key=="bar"][_key=="baz"][_key=="seg"]'
+  )
+  t.end()
+})
 
-    test('throws on unrecognized segment types', () => {
-      expect(() => toString([{foo: 'bar'}])).toThrowError(
-        'Unsupported path segment `{"foo":"bar"}`'
-      )
-    })
-  })
+test('toString: throws on unrecognized segment types', t => {
+  t.throws(() => toString([{foo: 'bar'}]), 'Unsupported path segment `{"foo":"bar"}`')
+  t.end()
+})
 
-  describe('get', () => {
-    test('throws on non-array/non-string path', () => {
-      expect(() => get(srcObject, null)).toThrowError('Path must be an array or a string')
-      expect(() => get(srcObject, 13)).toThrowError('Path must be an array or a string')
-      expect(() => get(srcObject, false)).toThrowError('Path must be an array or a string')
-      expect(() => get(srcObject, true)).toThrowError('Path must be an array or a string')
-    })
+test('get: throws on non-array/non-string path', t => {
+  t.throws(() => get(srcObject, null), 'Path must be an array or a string')
+  t.throws(() => get(srcObject, 13), 'Path must be an array or a string')
+  t.throws(() => get(srcObject, false), 'Path must be an array or a string')
+  t.throws(() => get(srcObject, true), 'Path must be an array or a string')
+  t.end()
+})
 
-    test('can get simple props', () => {
-      expect(get(srcObject, 'title')).toBe(srcObject.title)
-      expect(get(srcObject, ['title'])).toBe(srcObject.title)
-    })
+test('get: can get simple props', t => {
+  t.strictDeepEqual(get(srcObject, 'title'), srcObject.title)
+  t.strictDeepEqual(get(srcObject, ['title']), srcObject.title)
+  t.end()
+})
 
-    test('can pass default value', () => {
-      const defaultVal = Math.random()
-      expect(get(srcObject, 'notSet', defaultVal)).toBe(defaultVal)
-      expect(get(srcObject, ['notSet'], defaultVal)).toBe(defaultVal)
-    })
+test('get: can pass default value', t => {
+  const defaultVal = Math.random()
+  t.strictDeepEqual(get(srcObject, 'notSet', defaultVal), defaultVal)
+  t.strictDeepEqual(get(srcObject, ['notSet'], defaultVal), defaultVal)
+  t.end()
+})
 
-    test('can use array indexes', () => {
-      expect(get(srcObject, ['body', 0])).toBe(srcObject.body[0])
-      expect(get(srcObject, ['body', 1, 'children', 0])).toBe(srcObject.body[1].children[0])
-      expect(get(srcObject, ['multiDimensional', 1, 1])).toBe(srcObject.multiDimensional[1][1])
-    })
+test('get: can use array indexes', t => {
+  t.strictDeepEqual(get(srcObject, ['body', 0]), srcObject.body[0])
+  t.strictDeepEqual(get(srcObject, ['body', 1, 'children', 0]), srcObject.body[1].children[0])
+  t.strictDeepEqual(get(srcObject, ['multiDimensional', 1, 1]), srcObject.multiDimensional[1][1])
+  t.end()
+})
 
-    test('can use key lookup', () => {
-      expect(get(srcObject, ['body', {_key: 'bar'}])).toBe(srcObject.body[1])
-      expect(get(srcObject, ['body', {_key: 'bar'}, 'children', {_key: 'child1'}])).toBe(
-        srcObject.body[1].children[0]
-      )
-    })
+test('get: can use key lookup', t => {
+  t.strictDeepEqual(get(srcObject, ['body', {_key: 'bar'}]), srcObject.body[1])
+  t.strictDeepEqual(
+    get(srcObject, ['body', {_key: 'bar'}, 'children', {_key: 'child1'}]),
+    srcObject.body[1].children[0]
+  )
+  t.end()
+})
 
-    test('falls back to default value on array index at non-array', () => {
-      const defaultVal = {}
-      expect(get(srcObject, ['title', 1], defaultVal)).toBe(defaultVal)
-    })
+test('get: falls back to default value on array index at non-array', t => {
+  const defaultVal = {}
+  t.strictDeepEqual(get(srcObject, ['title', 1], defaultVal), defaultVal)
+  t.end()
+})
 
-    test('falls back to default value on property lookup at non-object', () => {
-      const defaultVal = {}
-      expect(get(srcObject, ['title', 'bar'], defaultVal)).toBe(defaultVal)
-      expect(get(srcObject, ['multiDimensional', 'bar'], defaultVal)).toBe(defaultVal)
-      expect(get(srcObject, ['nullVal', 'bar'], defaultVal)).toBe(defaultVal)
-    })
+test('get: falls back to default value on property lookup at non-object', t => {
+  const defaultVal = {}
+  t.strictDeepEqual(get(srcObject, ['title', 'bar'], defaultVal), defaultVal)
+  t.strictDeepEqual(get(srcObject, ['multiDimensional', 'bar'], defaultVal), defaultVal)
+  t.strictDeepEqual(get(srcObject, ['nullVal', 'bar'], defaultVal), defaultVal)
+  t.end()
+})
 
-    test('falls back to default value on key lookup at non-array', () => {
-      const defaultVal = {}
-      expect(get(srcObject, ['nullVal', {_key: 'abc'}], defaultVal)).toBe(defaultVal)
-    })
+test('get: falls back to default value on key lookup at non-array', t => {
+  const defaultVal = {}
+  t.strictDeepEqual(get(srcObject, ['nullVal', {_key: 'abc'}], defaultVal), defaultVal)
+  t.end()
+})
 
-    test('can get numbered key from object', () => {
-      expect(get(srcObject, 'nested.0')).toBe('Zero-Key')
-      expect(get(srcObject, ['nested', '0'])).toBe('Zero-Key')
-    })
-  })
+test('get: can get numbered key from object', t => {
+  t.strictDeepEqual(get(srcObject, 'nested.0'), 'Zero-Key')
+  t.strictDeepEqual(get(srcObject, ['nested', '0']), 'Zero-Key')
+  t.end()
 })
