@@ -43,7 +43,8 @@ export default {
     {
       name: 'coverImage',
       title: 'Cover Image',
-      type: 'image'
+      type: 'image',
+      options: {hotspot: true}
     },
     {
       name: 'publicationYear',
